Skip existing registration check to bypass database query issue
Temporarily removing the query to check for existing registrations since
that's where the 404 error is occurring. Will always create new registrations.
This should allow us to test if the database is accessible for CREATE operations
even if QUERY operations are failing.

Once we confirm CREATE works, we can debug why QUERY is failing separately.